Muistot: Image handling, separated controllers
